+++ src/main/java/ladysnake/effective/client/EffectiveConfig.java
@@ -25,7 +25,7 @@ public class EffectiveConfig extends MidnightConfig {
     public static int lapisBlockUpdateParticleChance = 10;
 
     @Entry(min = 0, max = 200)
-    public static int flowingWaterSplashingDesity = 50;
+    public static int flowingWaterSplashingDensity = 50;
 
     @Entry(min = 0, max = 10)
     public static int rainRippleDensity = 1;
@@ -38,4 +38,10 @@ public class EffectiveConfig extends MidnightConfig {
 
     @Entry
     public static boolean glowsquidHypnotizeAttractCursor = true;
+
+    @Entry
+    public static boolean enableAllayTrails = true;
+
+    @Entry(min = 0, max = 100)
+    public static int allayTwinkleDensity = 5;
 }