+++ src/main/java/ladysnake/effective/mixin/AbstractBlockMixin.java
@@ -1,6 +1,5 @@
 package ladysnake.effective.mixin;
 
-import ladysnake.effective.client.Effective;
 import ladysnake.effective.client.EffectiveConfig;
 import ladysnake.effective.client.world.WaterfallCloudGenerators;
 import net.minecraft.block.AbstractBlock;