+++ src/main/java/ladysnake/effective/mixin/WaterFluidMixin.java
@@ -9,7 +9,6 @@ import net.minecraft.particle.ParticleTypes;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.Vec3d;
-import net.minecraft.world.LightType;
 import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.BiomeKeys;
@@ -23,7 +22,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class WaterFluidMixin {
     @Unique
     private static boolean shouldSplash(World world, BlockPos pos) {
-        if (EffectiveConfig.flowingWaterSplashingDesity > 0) {
+        if (EffectiveConfig.flowingWaterSplashingDensity > 0) {
             FluidState fluidState = world.getFluidState(pos);
             if (!fluidState.isStill() & fluidState.getHeight() >= 0.77) {
                 BlockPos.Mutable mutable = new BlockPos.Mutable();
@@ -51,7 +50,7 @@ public class WaterFluidMixin {
         // flowing water splashes
         if (shouldSplash(world, pos.up())) {
             Vec3d vec3d = state.getVelocity(world, pos);
-            for (int i = 0; i <= random.nextInt(EffectiveConfig.flowingWaterSplashingDesity); i++) {
+            for (int i = 0; i <= random.nextInt(EffectiveConfig.flowingWaterSplashingDensity); i++) {
                 world.addParticle(ParticleTypes.SPLASH, pos.getX() + .5 + random.nextGaussian() / 2f, pos.getY() + 1 + random.nextFloat(), pos.getZ() + .5 + random.nextGaussian() / 2f, vec3d.getX() * random.nextFloat(), random.nextFloat() / 10f, vec3d.getZ() * random.nextFloat());
             }
         }